Note our controller isn’t actually doing anything yet, that’s okay.
The error we should be getting now is that we’re basically not returning
JSON. 

The strict TDD thing to do here (Remember: Red-Green-Refactor) 
would be to have our controller Return a static JSON string, then in 
the Refactor step, refactor our code to actually call a repository.

For the sake of brevity, i’m going to combine the two steps.