Once again the compiler will yet at you because of hanging references!
The compiler in this case is going to keep us in a “Red” state until 
We’ve implemented both the Crud Repository as well as the User Class. 

This is where statically typed languages a little awkward to test. In more
Dynamic languages (like Ruby, Python, and Javascript) we can leave all 
Hanging references in and run the code because for the most part those
Languages don’t care about those references until runtime.